fix: remove 'serial' python package; namespace conflict
Bug where having `serial` and `pyserial` in requirements results in a
namespace collision, and the error `AttributeError: module 'serial' has
no attribute 'Serial'` as the `serial` package is being referenced.

Confirmed that the package we actually want is `pyserial`, and removing
the `serial` package from the environment (and requirements) resolves
the issue.